Create an Erlang distribution

The Erlang distribution is a two-parameter family of continuous probability distributions with support \(x \in [0,\infty)\). The two parameters are a positive integer shape parameter \(k\) and a positive real rate parameter \(\lambda\). The Erlang distribution with shape parameter \(k = 1\) simplifies to the exponential distribution, and it is a special case of the gamma distribution. It corresponds to a sum of \(k\) independent exponential variables with mean \(1 / \lambda\) each.

Erlang(k, lambda)

Arguments

k

The shape parameter. Can be any positive integer number.

lambda

The rate parameter. Can be any positive number.

Value

An Erlang object.

See also

Examples

set.seed(27)

X <- Erlang(5, 2)
X
#> Erlang distribution (k = 5, lambda = 2)

random(X, 10)
#>  [1] 2.6290709 6.5928809 0.8851218 3.1279162 1.9834490 3.6187867 4.4790126
#>  [8] 1.6835867 3.5586337 4.8770382

pdf(X, 2)
#> [1] 0.3907336
log_pdf(X, 2)
#> [1] -0.9397292

cdf(X, 4)
#> [1] 0.9003676
quantile(X, 0.7)
#> [1] 2.945181

cdf(X, quantile(X, 0.7))
#> [1] 0.7
quantile(X, cdf(X, 7))
#> [1] 7