Create a Continuous Uniform distribution

A distribution with constant density on an interval. The continuous analogue to the Categorical() distribution.

Usage

Uniform(a = 0, b = 1)

Arguments

a

The a parameter. a can be any value in the set of real numbers. Defaults to 0.

b

The a parameter. b can be any value in the set of real numbers. It should be strictly bigger than a, but if is not, the order of the parameters is inverted. Defaults to 1.

Value

A Uniform object.

See also

Examples

set.seed(27)

X <- Uniform(1, 2)
X
#> [1] "Uniform(a = 1, b = 2)"

random(X, 10)
#>  [1] 1.971750 1.083758 1.873870 1.329231 1.222276 1.401648 1.072499 1.002450
#>  [9] 1.137094 1.191909

pdf(X, 0.7)
#> [1] 0
log_pdf(X, 0.7)
#> [1] -Inf

cdf(X, 0.7)
#> [1] 0
quantile(X, 0.7)
#> [1] 1.7

cdf(X, quantile(X, 0.7))
#> [1] 0.7
quantile(X, cdf(X, 0.7))
#> [1] 1