Evaluate the cumulative distribution function of a Beta distribution

Evaluate the cumulative distribution function of a Beta distribution

# S3 method for Beta
cdf(d, x, ...)

Arguments

d
A Beta object created by a call to Beta().

x
A vector of elements whose cumulative probabilities you would like to determine given the distribution d.

...
Unused. Unevaluated arguments will generate a warning to catch mispellings or other possible errors.

Value

A vector of probabilities, one for each element of x.

Examples

set.seed(27)

X <- Beta(1, 2)
X

#> Beta distribution (alpha = 1, beta = 2)

random(X, 10)

#>  [1] 0.014327255 0.067309943 0.636292291 0.864804440 0.758869543 0.237550867
#>  [7] 0.330895959 0.065843704 0.008265406 0.254705779

pdf(X, 0.7)

#> [1] 0.6

log_pdf(X, 0.7)

#> [1] -0.5108256

cdf(X, 0.7)

#> [1] 0.91

quantile(X, 0.7)

#> [1] 0.4522774

cdf(X, quantile(X, 0.7))

#> [1] 0.7

quantile(X, cdf(X, 0.7))

#> [1] 0.7