Determine quantiles of an Exponential distribution

quantile() is the inverse of cdf().

# S3 method for Exponential
quantile(x, probs, drop = TRUE, ...)

Arguments

x

An Exponential object created by a call to Exponential().

probs

A vector of probabilities.

drop

logical. Should the result be simplified to a vector if possible?

...

Arguments to be passed to qexp. Unevaluated arguments will generate a warning to catch mispellings or other possible errors.

Value

In case of a single distribution object, either a numeric vector of length probs (if drop = TRUE, default) or a matrix with length(probs) columns (if drop = FALSE). In case of a vectorized distribution object, a matrix with length(probs) columns containing all possible combinations.

Examples

set.seed(27)

X <- Exponential(5)
X
#> [1] "Exponential distribution (rate = 5)"

mean(X)
#> [1] 0.2
variance(X)
#> [1] 25
skewness(X)
#> [1] 2
kurtosis(X)
#> [1] 6

random(X, 10)
#>        r_1        r_2        r_3        r_4        r_5        r_6        r_7 
#> 0.01161126 0.28730930 1.15993941 0.29660927 0.38431337 0.04643808 0.06969554 
#>        r_8        r_9       r_10 
#> 0.10900366 0.50608948 0.03759968 

pdf(X, 2)
#> [1] 0.0002269996
log_pdf(X, 2)
#> [1] -8.390562

cdf(X, 4)
#> [1] 1
quantile(X, 0.7)
#> [1] 0.2407946

cdf(X, quantile(X, 0.7))
#> [1] 0.7
quantile(X, cdf(X, 7))
#> [1] 6.989008