Determine quantiles of a GEV distribution

quantile() is the inverse of cdf().

# S3 method for GEV
quantile(d, p, ...)

Arguments

d
A GEV object created by a call to GEV().

p
A vector of probabilities.

...
Unused. Unevaluated arguments will generate a warning to catch mispellings or other possible errors.

Value

A vector of quantiles, one for each element of p.

Examples

set.seed(27)

X <- GEV(1, 2, 0.1)
X
#> GEV distribution (mu = 1, sigma = 2, xi = 0.1)

random(X, 10)
#>  [1]  9.53039102 -0.73633998  5.43730770  0.79059280  0.20038342  1.18468635
#>  [7] -0.83938790 -2.28404509 -0.32725032  0.02226797

pdf(X, 0.7)
#> [1] 0.1845098
log_pdf(X, 0.7)
#> [1] -1.690052

cdf(X, 0.7)
#> [1] 0.3124986
quantile(X, 0.7)
#> [1] 3.171891

cdf(X, quantile(X, 0.7))
#> [1] 0.7
quantile(X, cdf(X, 0.7))
#> [1] 0.7