Determine quantiles of a NegativeBinomial distribution

Determine quantiles of a NegativeBinomial distribution

# S3 method for NegativeBinomial
quantile(x, probs, drop = TRUE, ...)

Arguments

x

A NegativeBinomial object created by a call to NegativeBinomial().

probs

A vector of probabilities.

drop

logical. Should the result be simplified to a vector if possible?

...

Arguments to be passed to qnbinom. Unevaluated arguments will generate a warning to catch mispellings or other possible errors.

Value

In case of a single distribution object, either a numeric vector of length probs (if drop = TRUE, default) or a matrix with length(probs) columns (if drop = FALSE). In case of a vectorized distribution object, a matrix with length(probs) columns containing all possible combinations.

See also

Examples

set.seed(27)

X <- NegativeBinomial(10, 0.3)
X
#> [1] "NegativeBinomial distribution (size = 10, p = 0.3)"

random(X, 10)
#>  r_1  r_2  r_3  r_4  r_5  r_6  r_7  r_8  r_9 r_10 
#>   45   25   27   41   13   11   16   22   12   23 

pdf(X, 2)
#> [1] 0.0001591371
log_pdf(X, 2)
#> [1] -8.745745

cdf(X, 4)
#> [1] 0.001665663
quantile(X, 0.7)
#> [1] 27