Determine quantiles of a continuous Uniform distribution

quantile() is the inverse of cdf().

# S3 method for Uniform
quantile(d, p, ...)

Arguments

d
A Uniform object created by a call to Uniform().

p
A vector of probabilites.

...
Unused. Unevaluated arguments will generate a warning to catch mispellings or other possible errors.

Value

A vector of quantiles, one for each element of p.

Examples

set.seed(27)

X <- Uniform(1, 2)
X

#> Continuous Uniform distribution (a = 1, b = 2)

random(X, 10)

#>  [1] 1.971750 1.083758 1.873870 1.329231 1.222276 1.401648 1.072499 1.002450
#>  [9] 1.137094 1.191909

pdf(X, 0.7)

#> [1] 0

log_pdf(X, 0.7)

#> [1] -Inf

cdf(X, 0.7)

#> [1] 0

quantile(X, 0.7)

#> [1] 1.7

cdf(X, quantile(X, 0.7))

#> [1] 0.7

quantile(X, cdf(X, 0.7))

#> [1] 1