Draw a random sample from a Logistic distribution

Draw a random sample from a Logistic distribution

# S3 method for Logistic
random(d, n = 1L, ...)

Arguments

d
A Logistic object created by a call to Logistic().

n
The number of samples to draw. Defaults to 1L.

...
Unused. Unevaluated arguments will generate a warning to catch mispellings or other possible errors.

Value

An integer vector of length n.

See also

Examples

set.seed(27)

X <- Logistic(2, 4)
X
#> Logistic distribution (location = 2, scale = 4)

random(X, 10)
#>  [1]  16.1520541  -7.5694209   9.7424712  -0.8466541  -3.0098187   0.4055911
#>  [7]  -8.1957130 -22.0364748  -5.3585558  -3.7506119

pdf(X, 2)
#> [1] 0.0625
log_pdf(X, 2)
#> [1] -2.772589

cdf(X, 4)
#> [1] 0.6224593
quantile(X, 0.7)
#> [1] 5.389191