Return the support of the Beta distribution

Return the support of the Beta distribution

# S3 method for Beta
support(d)

Arguments

d
An Beta object created by a call to Beta().

Value

A vector of length 2 with the minimum and maximum value of the support.