distributions3: Probability Distributions as S3 Objects

Tools to create and manipulate probability distributions using S3. Generics pdf(), cdf(), quantile(), and random() provide replacements for base R's d/p/q/r style functions. Functions and arguments have been named carefully to minimize confusion for students in intro stats courses. The documentation for each distribution contains detailed mathematical notes.

See also

Author

Maintainer: Alex Hayes alexpghayes@gmail.com (ORCID)

Authors:

Ralph Moller-Trane

Daniel Jordan dandermotj@gmail.com

Paul Northrop p.northrop@ucl.ac.uk

Moritz N. Lang moritz.n.lang@gmail.com (ORCID)

Achim Zeileis Achim.Zeileis@R-project.org (ORCID)

Other contributors:

Emil Hvitfeldt emilhhvitfeldt@gmail.com (ORCID) [contributor]

Bruna Wundervald brunadaviesw@gmail.com [contributor]

Alessandro Gasparini alessandro.gasparini@ki.se [contributor]